+++ examples/data-sources/airlock_cloud_groups/data-source.tf
@@ -1,0 +1,1 @@
+data "airlock_cloud_groups" "all" {}